requirements.txt
PyMuPDF==1.26.4
langchain==0.3.27
langchain_community==0.3.29
langchain_core==0.3.76
langchain_google_genai==2.1.10
python-dotenv==1.1.1
PyYAML==6.0.2
streamlit==1.41.1
chromadb==0.6.3
